pyproject.toml
[tool.poetry]
name = "pyobs-qhyccd"
version = "1.1.0"
description = "pyobs module for QHYCCD cameras"
authors = ["Tim-Oliver Husser <thusser@uni-goettingen.de>"]
license = "MIT"
include = ['lib']

[tool.poetry.build]
script = "build.py"
generate-setup-file = false

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
astropy = "^5.0"
numpy = "^1.21"
pyobs-core = "^1.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
pre-commit = "^2.16"
sphinx-rtd-theme = "^1.0"
Sphinx = "^4.4"
cython = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0", "Cython>=3.0.0", "numpy>=1.21.0", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
